// src/app/shared/model/pIDetailsModel.ts
export interface PIDetails {
    requestNo: string;
    projectCode: string;
    nicsipino: string;
    pidate: string;
    piamount: string;
    fileContent: string;
    docType: string;
    modifyBy: string;
    modifyOn: string;
    fileContentCover: string;
    fileContentProposal: string
    docTypeCover: string;
    docTypeProposal: string;
    //docFileName: string;
    docContentType: string;
    docFileNameUploadPI: string,
    docFileNameCoverLetter: string,
    docFileNameProposal: string,
}